# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BrandFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/generate:
    post:
      operationId: generateContent
      tags: [ai]
      summary: Generate social media post content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          description: Generated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResult"
  /ai/hashtags:
    post:
      operationId: generateHashtags
      tags: [ai]
      summary: Generate hashtags for a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HashtagRequest"
      responses:
        "200":
          description: Generated hashtags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HashtagResult"
  /ai/reply:
    post:
      operationId: generateReply
      tags: [ai]
      summary: Generate an auto-reply in brand voice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyRequest"
      responses:
        "200":
          description: Generated reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResult"
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Chat with the AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResult"

  # ─── Analytics ────────────────────────────────────────────────────────────
...
  /comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments and mentions
      parameters:
        - name: platform
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
  /comments/{id}/reply:
    post:
      operationId: replyToComment
      tags: [comments]
      summary: Reply to a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentReplyInput"
      responses:
        "200":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}/status:
    patch:
      operationId: updateCommentStatus
      tags: [comments]
      summary: Update comment status (resolve, escalate, spam)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # ─── Brand ───────────────────────────────────────────────────────────────
...
  /knowledge:
    get:
      operationId: listKnowledge
      tags: [knowledge]
      summary: List knowledge base items
      responses:
        "200":
          description: Knowledge items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeItem"
    post:
      operationId: createKnowledge
      tags: [knowledge]
      summary: Add a knowledge base item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItem"
  /knowledge/{id}:
    delete:
      operationId: deleteKnowledge
      tags: [knowledge]
      summary: Delete a knowledge item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── Team ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    GenerateRequest:
      type: object
      required: [topic, platform]
      properties:
        topic:
          type: string
        platform:
          type: string
          enum: [twitter, linkedin, instagram, facebook]
        tone:
          type: string
          enum: [professional, funny, casual, luxury, startup, corporate]
        includeHashtags:
          type: boolean
        includeEmojis:
          type: boolean
        variations:
          type: number
    GenerateResult:
      type: object
      required: [variations]
      properties:
        variations:
          type: array
          items:
            $ref: "#/components/schemas/GeneratedPost"
    GeneratedPost:
      type: object
      required: [content]
      properties:
        content:
          type: string
        hashtags:
          type: string
        cta:
          type: string
    HashtagRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
        platform:
          type: string
        count:
          type: number
    HashtagResult:
      type: object
      required: [hashtags]
      properties:
        hashtags:
          type: array
          items:
            type: string
    ReplyRequest:
      type: object
      required: [commentText, platform]
      properties:
        commentText:
          type: string
        platform:
          type: string
        context:
          type: string
    ReplyResult:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
        sentiment:
          type: string
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
        conversationId:
          type: ["string", "null"]
    ChatResult:
      type: object
      required: [response]
      properties:
        response:
          type: string
        conversationId:
          type: string

    # ─── Analytics ───────────────────────────────────────────────────────────
...
    Comment:
      type: object
      required: [id, authorName, content, platform, status, createdAt]
      properties:
        id:
          type: number
        postId:
          type: ["number", "null"]
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        content:
          type: string
        platform:
          type: string
        status:
          type: string
          enum: [pending, replied, resolved, escalated, spam]
        sentiment:
          type: ["string", "null"]
        reply:
          type: ["string", "null"]
        createdAt:
          type: string
    CommentReplyInput:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
    CommentStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, replied, resolved, escalated, spam]

    # ─── Brand ───────────────────────────────────────────────────────────────
...
    KnowledgeItem:
      type: object
      required: [id, title, type, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        content:
          type: ["string", "null"]
        type:
          type: string
          enum: [document, product, faq, post_example, guideline]
        fileUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    KnowledgeInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [document, product, faq, post_example, guideline]
        fileUrl:
          type: string

    # ─── Team ────────────────────────────────────────────────────────────────
...